일단 올려

+++ ZOOTOPIA/src/main/java/com/aloha/zootopia/domain/Comment.java
@@ -13,17 +13,13 @@ import java.util.Date;
 @Builder
 public class Comment {
 
-    private Integer commentId;   
+    private int commentId;   
     private String content;      
-
     private Date createdAt;      
     private Date updatedAt;      
     private Boolean isDeleted;   
-
-    private Integer userId;     
-    private Integer postId;      
-
-    
+    private Long userId;     
+    private int postId;      
     private String nickname;     
     private String profileImg;   
 }
